LayoutTests/fast/regions/selection/crash-deselect-expected.txt

CONSOLE MESSAGE: line 66: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.

Errlog selectionTest: HierarchyRequestError: HierarchyRequestError: DOM Exception 3 : index 1 
Errlog webtest_fn_24: 
